src/requirements.txt
django~=3.2.10
djangorestframework~=3.13.1
psycopg2-binary~=2.9.2
python-decouple~=3.5
uwsgi~=2.0.20
django-redis-cache~=3.0.0
pandas~=1.3.5
